doc: Fix dangerous dcim-transfer.service example
The sample systemd service provided in the upload-from-your-camera guide
contains a one-line shell script that can inadvertently wipe the SD card
after failing to import photos off of it.

(This happened to me yesterday, and it was not fun,
though it's possible to recover files with photorec.)

This commit fixes the script to prevent such accidental deletions.