bfx: Filter out names of deleted files at import
The debouncer accumulates a list of
names of files that appear in the inbox within n seconds of each other.
Between the time the debouncer is first triggered
and the time its code is executed,
it's possible that some of those files could have disappeared--
but the import logic will still run for _all_ accumulated filenames,
including the deleted ones.

This commit fixes that
by filtering out the names of those deleted files
before running the import logic.