rfg: Use photein without spawning a subprocess
Originally, xferase imported files
by invoking its sister utility, photein, as a subprocess
(via the `Kernel#system` method).

This commit unifies all the program logic into a single Ruby process
by invoking Photein as a Ruby library rather than a CLI executable.

The principal benefit of this change is
it guarantees that Xferase will use the proper version of Photein,
as specified in the Gemfile.
(In the past, using `Kernel#system` meant that it would run
whichever version of Photein was available on the user's $PATH.)

Another benefit of this approach is that
all photein log output can be delegated directly to the Xferase::Logger,
ensuring that Xferase and Photein are both outputting log messages
at the same level of verbosity.
(In the past, this was achieved by
conditionally/explicitly specifying the `--verbose` flag
when invoking photein via the `Kernel#system` method.)