ftr: Simplify interface (remove staging/grace opts)
This commit leverages a new feature
available in the upstream/sister utility Photein
to encode/import files to both desktop and web libraries in parallel
(i.e., using multi-threaded logic),
thus eliminating the need for a staging directory entirely.

This commit also removes the `--grace-period` CLI option
with the rationale that the slow upload/synchronization of files
from a remote device (e.g., a cell phone) to the host system
provides a sufficient "grace period" on its own.